backend/api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.4
xgboost==2.0.2
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
scikit-learn